refactor(mocks): extract agents module from mockApi (#991 M3)
Move template/skill/allocation state and helpers (incl. sameContent) and the 17 listAgentTemplates…setTemplateSkills methods into agents.ts as agentsApi; index.ts spreads it, imports templates/LEAD_NAME_RE for the runs own_agents read, and re-exports sameContent for the drift-contract test.
